APIMF-2944: define collectors interface ensuring only one traversal is made
