+++ amf-webapi/shared/src/main/scala/amf/plugins/document/webapi/validation/UnitPayloadsValidation.scala
@@ -4,19 +4,17 @@ import amf.core.benchmark.ExecutionLog
 import amf.core.metamodel.document.PayloadFragmentModel
 import amf.core.model.document.BaseUnit
 import amf.core.model.domain.{ArrayNode, DataNode, ObjectNode}
-import amf.core.remote.Platform
 import amf.core.validation.{AMFValidationReport, AMFValidationResult, SeverityLevels, ValidationCandidate}
 import amf.internal.environment.Environment
-import amf.plugins.domain.shapes.validation.{PayloadValidationPluginsHandler}
+import amf.plugins.document.webapi.validation.collector.{CollectorsRunner, ValidationCandidateCollector}
+import amf.plugins.domain.shapes.validation.PayloadValidationPluginsHandler
 import amf.validations.PayloadValidations
 
 import scala.concurrent.{ExecutionContext, Future}
 
-case class UnitPayloadsValidation(baseUnit: BaseUnit) {
+case class UnitPayloadsValidation(baseUnit: BaseUnit, collectors: Seq[ValidationCandidateCollector]) {
 
-  val candidates: Seq[ValidationCandidate] = PayloadsInApiCollector(baseUnit) ++
-    ShapeFacetsCandidatesCollector(baseUnit) ++
-    ExtensionsCandidatesCollector(baseUnit)
+  val candidates: Seq[ValidationCandidate] = CollectorsRunner(collectors).traverse(baseUnit).toSeq
 
   val index: DataNodeIndex = {
     val nodes = getCandidateNodes
