+++ amf-webapi/shared/src/main/scala/amf/plugins/document/webapi/validation/runner/steps/ExamplesValidationStep.scala
@@ -2,6 +2,12 @@ package amf.plugins.document.webapi.validation.runner.steps
 
 import amf.core.validation.{AMFValidationReport, AMFValidationResult, ShaclReportAdaptation}
 import amf.plugins.document.webapi.validation.UnitPayloadsValidation
+import amf.plugins.document.webapi.validation.collector.{
+  EnumInShapesCollector,
+  ExtensionsCollector,
+  PayloadsInApiCollector,
+  ShapeFacetsCollector
+}
 import amf.plugins.document.webapi.validation.runner.ValidationContext
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -11,7 +17,8 @@ case class ExamplesValidationStep(override val validationContext: ValidationCont
     with ShaclReportAdaptation {
 
   override protected def validate()(implicit executionContext: ExecutionContext): Future[AMFValidationReport] = {
-    UnitPayloadsValidation(validationContext.baseUnit)
+    val collectors = Seq(PayloadsInApiCollector, EnumInShapesCollector, ShapeFacetsCollector, ExtensionsCollector)
+    UnitPayloadsValidation(validationContext.baseUnit, collectors)
       .validate(validationContext.env)
       .map { results =>
         val mappedSeverityResults = results.flatMap { buildValidationWithCustomLevelForProfile }
